+++ bricks/_common/micropython.c
@@ -292,6 +292,9 @@ void pbsys_main_run_program(pbsys_main_program_t *program) {
     }
     #endif // PYBRICKS_OPT_COMPILER
 
+    // Ensure all buffered output has been sent over the air.
+    mp_hal_stdout_tx_flush();
+
     // Clean up non-MicroPython resources used by the pybricks package.
     pb_package_pybricks_deinit();